requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
pydantic
psycopg2-binary
python-dotenv